src/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
